@@ -9,59 +9,56 @@
     "sitemap": "cross-env NODE_OPTIONS='--experimental-json-modules' node ./scripts/generate-sitemap.mjs",
     "rss": "cross-env NODE_OPTIONS='--experimental-json-modules' node ./scripts/generate-rss.mjs",
     "serve": "next start",
-    "analyze": "cross-env ANALYZE=true next build",
     "lint": "next lint --fix --dir pages --dir components --dir lib --dir layouts --dir scripts",
     "prepare": "husky install"
+  },
+  "engines": {
+    "node": "18"
   },
   "dependencies": {
     "@dlarroder/playground": "1.2.2",
     "@fontsource/mukta": "4.5.10",
     "@next/font": "13.0.1",
+    "@studio-freight/react-lenis": "^0.0.21",
     "@vercel/analytics": "0.1.1",
     "classnames": "2.3.2",
     "cmdk": "0.1.20",
-    "contentlayer": "0.2.9",
+    "contentlayer": "0.3.2",
     "esbuild": "0.16.12",
     "framer-motion": "10.12.16",
     "github-slugger": "1.4.0",
     "gsap": "3.12.1",
-    "locomotive-scroll": "4.1.4",
     "logrocket": "3.0.1",
     "mdx-bundler": "9.0.1",
-    "next": "13.1.1",
-    "next-contentlayer": "0.2.9",
+    "next": "^13.4.6",
+    "next-contentlayer": "0.3.2",
     "next-themes": "0.2.1",
-    "nprogress": "0.2.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "4.6.0",
-    "react-locomotive-scroll": "0.2.2",
     "reading-time": "1.5.0",
     "rehype-autolink-headings": "6.1.1",
     "rehype-prism-plus": "1.5.0",
     "rehype-slug": "5.1.0",
     "remark": "14.0.2",
     "sharp": "0.31.3",
-    "swr": "2.0.0",
     "use-breakpoint": "3.1.1"
   },
   "devDependencies": {
-    "@next/bundle-analyzer": "13.1.1",
     "@svgr/webpack": "6.5.1",
     "@tailwindcss/forms": "0.5.3",
     "@tailwindcss/typography": "0.5.8",
     "@types/github-slugger": "1.3.0",
-    "@types/locomotive-scroll": "4.1.0",
     "@types/node": "18.11.18",
     "@types/nprogress": "0.2.0",
-    "@types/react": "18.0.26",
+    "@types/react": "18.2.8",
     "@typescript-eslint/eslint-plugin": "5.40.1",
     "@typescript-eslint/parser": "5.40.1",
     "autoprefixer": "10.4.13",
     "cross-env": "7.0.3",
     "dedent": "0.7.0",
     "eslint": "8.31.0",
-    "eslint-config-next": "13.1.1",
+    "eslint-config-next": "^13.4.6",
     "eslint-config-prettier": "8.6.0",
     "eslint-plugin-jsx-a11y": "6.6.1",
     "eslint-plugin-prettier": "4.2.1",
@@ -75,7 +72,7 @@
     "prettier-plugin-tailwindcss": "0.2.1",
     "tailwind-scrollbar": "2.0.1",
     "tailwindcss": "3.2.4",
-    "typescript": "4.9.4"
+    "typescript": "5.1.3"
   },
   "lint-staged": {
     "*.+(js|jsx|ts|tsx)": [
