@@ -1,5 +1,5 @@
-import NextImage, { ImageProps } from 'next/image'
+import NextImage, { ImageProps } from 'next/image';
 
-const Image = ({ ...rest }: ImageProps) => <NextImage {...rest} />
+const Image = ({ ...rest }: ImageProps) => <NextImage {...rest} />;
 
-export default Image
+export default Image;
