@@ -1,7 +1,7 @@
-import { ReactNode } from 'react'
+import { ReactNode } from 'react';
 
 interface Props {
-  children: ReactNode
+  children: ReactNode;
 }
 
 export default function PageTitle({ children }: Props) {
@@ -9,5 +9,5 @@ export default function PageTitle({ children }: Props) {
     <h1 className="text-2xl font-extrabold leading-9 tracking-tight text-white sm:text-4xl sm:leading-10 md:text-5xl md:leading-14">
       {children}
     </h1>
-  )
+  );
 }
