@@ -1,16 +1,16 @@
-import Link from '@/components/Link'
+import Link from 'next/link';
 
 interface Props {
-  totalPages: number
-  currentPage: number
+  totalPages: number;
+  currentPage: number;
 }
 
 export default function Pagination({ totalPages, currentPage }: Props) {
-  const prevPage = currentPage - 1 > 0
-  const nextPage = currentPage + 1 <= totalPages
+  const prevPage = currentPage - 1 > 0;
+  const nextPage = currentPage + 1 <= totalPages;
 
   return (
-    <div className="space-y-2 pt-6 pb-8 md:space-y-5">
+    <div className="space-y-2 md:space-y-5">
       <nav className="flex justify-between">
         {!prevPage && (
           <button className="cursor-auto disabled:opacity-50" disabled={!prevPage}>
@@ -18,7 +18,10 @@ export default function Pagination({ totalPages, currentPage }: Props) {
           </button>
         )}
         {prevPage && (
-          <Link href={currentPage - 1 === 1 ? `/blog/` : `/blog/page/${currentPage - 1}`}>
+          <Link
+            href={currentPage - 1 === 1 ? `/blog/` : `/blog/page/${currentPage - 1}`}
+            className="underline-magical"
+          >
             <button>Previous</button>
           </Link>
         )}
@@ -31,11 +34,11 @@ export default function Pagination({ totalPages, currentPage }: Props) {
           </button>
         )}
         {nextPage && (
-          <Link href={`/blog/page/${currentPage + 1}`}>
+          <Link href={`/blog/page/${currentPage + 1}`} className="underline-magical">
             <button>Next</button>
           </Link>
         )}
       </nav>
     </div>
-  )
+  );
 }
