@@ -1,12 +1,14 @@
-import Tag from '@/components/Tag'
-import { CoreContent } from '@/lib/utils/contentlayer'
-import type { Blog } from 'contentlayer/generated'
-import { motion } from 'framer-motion'
-import Link from 'next/link'
+'use client';
+
+import Tag from '@/components/Tag';
+import { CoreContent } from '@/lib/utils/contentlayer';
+import type { Blog } from 'contentlayer/generated';
+import { motion } from 'framer-motion';
+import Link from 'next/link';
 
 export interface PostCardProps {
-  posts: CoreContent<Blog>[]
-  showTags?: boolean
+  posts: CoreContent<Blog>[];
+  showTags?: boolean;
 }
 
 export default function PostCard({ posts, showTags = true }: PostCardProps) {
@@ -20,7 +22,7 @@ export default function PostCard({ posts, showTags = true }: PostCardProps) {
           animate={{ scale: 1, opacity: 1 }}
           transition={{ duration: 0.6, delay: index / 10 }}
         >
-          <Link href={`/blog/${slug}`} aria-label={`Read "${title}"`}>
+          <Link href={`/blog/${slug}`} aria-label={`Read "${title}"`} legacyBehavior>
             <article className="cursor-pointer gap-3 space-y-2 bg-opacity-20 py-5 xl:grid xl:grid-cols-4 xl:items-baseline xl:space-y-0">
               <div className="space-y-3 xl:col-span-4">
                 <span className="text-2xl font-bold leading-8 tracking-tight">
@@ -44,5 +46,5 @@ export default function PostCard({ posts, showTags = true }: PostCardProps) {
         </motion.li>
       ))}
     </ul>
-  )
+  );
 }
