@@ -1,19 +1,19 @@
-import { PostCardProps } from '@/components/PostCard/PostCard'
-import dynamic from 'next/dynamic'
+import { PostCardProps } from '@/components/PostCard/PostCard';
+import dynamic from 'next/dynamic';
 
 const PostCardComponent = dynamic(
   () => {
-    return import('./PostCard')
+    return import('./PostCard');
   },
   { ssr: false }
-)
+);
 
 const PostCard = ({ posts, showTags }: PostCardProps) => {
   return (
     <div id="comment">
       <PostCardComponent posts={posts} showTags={showTags} />
     </div>
-  )
-}
+  );
+};
 
-export default PostCard
+export default PostCard;
