@@ -1,12 +1,14 @@
-import { motion } from 'framer-motion'
-import gsap from 'gsap'
-import Image from 'next/image'
-import { useEffect, useRef } from 'react'
-import { Project, ProjectModal } from './types'
+'use client';
+
+import { motion } from 'framer-motion';
+import gsap from 'gsap';
+import Image from 'next/image';
+import { useEffect, useRef } from 'react';
+import { Project, ProjectModal } from './types';
 
 interface ProjectModalProps {
-  modal: ProjectModal
-  projects: Project[]
+  modal: ProjectModal;
+  projects: Project[];
 }
 
 const scaleAnimation = {
@@ -23,49 +25,51 @@ const scaleAnimation = {
     y: '-50%',
     transition: { duration: 0.4, ease: [0.32, 0, 0.67, 0] },
   },
-}
+};
 
 export default function ProjectPreview({ modal, projects }: ProjectModalProps) {
-  const { active, index } = modal
-  const modalContainer = useRef<HTMLDivElement>(null)
-  const cursor = useRef<HTMLDivElement>(null)
-  const cursorLabel = useRef<HTMLDivElement>(null)
+  const { active, index } = modal;
+  const modalContainer = useRef<HTMLDivElement>(null);
+  const cursor = useRef<HTMLDivElement>(null);
+  const cursorLabel = useRef<HTMLDivElement>(null);
 
   useEffect(() => {
-    // Move Container
-    const xMoveContainer = gsap.quickTo(modalContainer.current, 'left', {
-      duration: 0.8,
-      ease: 'power3',
-    })
-    const yMoveContainer = gsap.quickTo(modalContainer.current, 'top', {
-      duration: 0.8,
-      ease: 'power3',
-    })
+    if (window !== undefined) {
+      // Move Container
+      const xMoveContainer = gsap.quickTo(modalContainer.current, 'left', {
+        duration: 0.8,
+        ease: 'power3',
+      });
+      const yMoveContainer = gsap.quickTo(modalContainer.current, 'top', {
+        duration: 0.8,
+        ease: 'power3',
+      });
 
-    // Move cursor
-    const xMoveCursor = gsap.quickTo(cursor.current, 'left', { duration: 0.5, ease: 'power3' })
-    const yMoveCursor = gsap.quickTo(cursor.current, 'top', { duration: 0.5, ease: 'power3' })
+      // Move cursor
+      const xMoveCursor = gsap.quickTo(cursor.current, 'left', { duration: 0.5, ease: 'power3' });
+      const yMoveCursor = gsap.quickTo(cursor.current, 'top', { duration: 0.5, ease: 'power3' });
 
-    // Move cursor label
-    const xMoveCursorLabel = gsap.quickTo(cursorLabel.current, 'left', {
-      duration: 0.45,
-      ease: 'power3',
-    })
-    const yMoveCursorLabel = gsap.quickTo(cursorLabel.current, 'top', {
-      duration: 0.45,
-      ease: 'power3',
-    })
+      // Move cursor label
+      const xMoveCursorLabel = gsap.quickTo(cursorLabel.current, 'left', {
+        duration: 0.45,
+        ease: 'power3',
+      });
+      const yMoveCursorLabel = gsap.quickTo(cursorLabel.current, 'top', {
+        duration: 0.45,
+        ease: 'power3',
+      });
 
-    window.addEventListener('mousemove', (e) => {
-      const { pageX, pageY } = e
-      xMoveContainer(pageX)
-      yMoveContainer(pageY)
-      xMoveCursor(pageX)
-      yMoveCursor(pageY)
-      xMoveCursorLabel(pageX)
-      yMoveCursorLabel(pageY)
-    })
-  }, [])
+      window.addEventListener('mousemove', (e) => {
+        const { pageX, pageY } = e;
+        xMoveContainer(pageX);
+        yMoveContainer(pageY);
+        xMoveCursor(pageX);
+        yMoveCursor(pageY);
+        xMoveCursorLabel(pageX);
+        yMoveCursorLabel(pageY);
+      });
+    }
+  }, []);
 
   return (
     <>
@@ -81,7 +85,7 @@ export default function ProjectPreview({ modal, projects }: ProjectModalProps) {
           style={{ top: index * -100 + '%', transition: 'top 0.6s cubic-bezier(0.76, 0, 0.24, 1)' }}
         >
           {projects.map((project, index) => {
-            const { src, color } = project
+            const { src, color } = project;
             return (
               <div
                 className="flex h-full w-full items-center justify-center"
@@ -96,7 +100,7 @@ export default function ProjectPreview({ modal, projects }: ProjectModalProps) {
                   alt="image"
                 />
               </div>
-            )
+            );
           })}
         </div>
       </motion.div>
@@ -117,5 +121,5 @@ export default function ProjectPreview({ modal, projects }: ProjectModalProps) {
         View
       </motion.div>
     </>
-  )
+  );
 }
