@@ -1,12 +1,12 @@
 export interface Project {
-  title: string
-  src: string
-  color: string
-  url: string
-  role: string
+  title: string;
+  src: string;
+  color: string;
+  url: string;
+  role: string;
 }
 
 export interface ProjectModal {
-  active: boolean
-  index: number
+  active: boolean;
+  index: number;
 }
