@@ -1,14 +1,14 @@
-import dynamic from 'next/dynamic'
+import dynamic from 'next/dynamic';
 
 const ScrollProgressBarComponent = dynamic(
   () => {
-    return import('./ScrollProgressBar')
+    return import('./ScrollProgressBar');
   },
   { ssr: false }
-)
+);
 
 const ScrollProgressBar = () => {
-  return <ScrollProgressBarComponent />
-}
+  return <ScrollProgressBarComponent />;
+};
 
-export default ScrollProgressBar
+export default ScrollProgressBar;
