@@ -1,8 +1,10 @@
-import { createContext } from 'react'
+'use client';
+
+import { createContext } from 'react';
 
 interface TileContextValue {
-  numOfPages: number
-  currentPage: number
+  numOfPages: number;
+  currentPage: number;
 }
 
-export const TileContext = createContext<TileContextValue>({ numOfPages: 0, currentPage: 0 })
+export const TileContext = createContext<TileContextValue>({ numOfPages: 0, currentPage: 0 });
