@@ -1,12 +1,12 @@
-import { ReactNode } from 'react'
+import { ReactNode } from 'react';
 
 export interface WorkProps {
-  children: ReactNode
-  progress: number
+  children: ReactNode;
+  progress: number;
 }
 
 export function WorkRight({ children, progress }: WorkProps) {
-  const translateY = Math.max(-50, -(progress - 0.5) * 50)
+  const translateY = Math.max(-50, -(progress - 0.5) * 50);
 
   return (
     <div
@@ -15,5 +15,5 @@ export function WorkRight({ children, progress }: WorkProps) {
     >
       <div className="w-full max-w-md px-10 pt-10 md:px-0 lg:pt-0">{children}</div>
     </div>
-  )
+  );
 }
