@@ -1,12 +1,12 @@
 export type WorkTile = {
-  title: string
-  description: string
+  title: string;
+  description: string;
   image: {
-    src: string
-    width: number
-    height: number
-  }
-}
+    src: string;
+    width: number;
+    height: number;
+  };
+};
 
 export const workTiles: WorkTile[] = [
   {
@@ -45,4 +45,4 @@ export const workTiles: WorkTile[] = [
       height: 717,
     },
   },
-]
+];
