@@ -1,7 +1,7 @@
-import { escape } from '@/lib/utils/htmlEscaper'
+import { escape } from '@/lib/utils/htmlEscaper';
 
-import siteMetadata from '@/data/siteMetadata'
-import { PostFrontMatter } from 'types/PostFrontMatter'
+import siteMetadata from '@/data/siteMetadata';
+import { PostFrontMatter } from 'types/PostFrontMatter';
 
 const generateRssItem = (post: PostFrontMatter) => `
   <item>
@@ -13,7 +13,7 @@ const generateRssItem = (post: PostFrontMatter) => `
     <author>${siteMetadata.email} (${siteMetadata.author})</author>
     ${post.tags && post.tags.map((t) => `<category>${t}</category>`).join('')}
   </item>
-`
+`;
 
 const generateRss = (posts: PostFrontMatter[], page = 'feed.xml') => `
   <rss version="2.0" xmlns:atom="http://www.w3.org/2005/Atom">
@@ -29,5 +29,5 @@ const generateRss = (posts: PostFrontMatter[], page = 'feed.xml') => `
       ${posts.map(generateRssItem).join('')}
     </channel>
   </rss>
-`
-export default generateRss
+`;
+export default generateRss;
