@@ -1,14 +1,14 @@
-import siteMetadata from '@/data/siteMetadata'
+import siteMetadata from '@/data/siteMetadata';
 
 const formatDate = (date: string) => {
   const options: Intl.DateTimeFormatOptions = {
     year: 'numeric',
     month: 'long',
     day: 'numeric',
-  }
-  const now = new Date(date).toLocaleDateString(siteMetadata.locale, options)
+  };
+  const now = new Date(date).toLocaleDateString(siteMetadata.locale, options);
 
-  return now
-}
+  return now;
+};
 
-export default formatDate
+export default formatDate;
