@@ -1,8 +1,8 @@
 // @ts-nocheck
-const { replace } = ''
+const { replace } = '';
 
 // escape
-const ca = /[&<>'"]/g
+const ca = /[&<>'"]/g;
 
 const esca = {
   '&': '&amp;',
@@ -10,8 +10,8 @@ const esca = {
   '>': '&gt;',
   "'": '&#39;',
   '"': '&quot;',
-}
-const pe = (m: keyof typeof esca) => esca[m]
+};
+const pe = (m: keyof typeof esca) => esca[m];
 
 /**
  * Safely escape HTML entities such as `&`, `<`, `>`, `"`, and `'`.
@@ -20,4 +20,4 @@ const pe = (m: keyof typeof esca) => esca[m]
  *  the input type is unexpected, except for boolean and numbers,
  *  converted as string.
  */
-export const escape = (es: string): string => replace.call(es, ca, pe)
+export const escape = (es: string): string => replace.call(es, ca, pe);
