@@ -1,27 +1,27 @@
-import { writeFileSync, mkdirSync } from 'fs'
-import path from 'path'
-import GithubSlugger from 'github-slugger'
-import { escape } from './htmlEscaper.mjs'
-import siteMetadata from '../data/siteMetadata.js'
-import { allBlogs } from '../.contentlayer/generated/index.mjs'
+import { writeFileSync, mkdirSync } from 'fs';
+import path from 'path';
+import GithubSlugger from 'github-slugger';
+import { escape } from './htmlEscaper.mjs';
+import siteMetadata from '../data/siteMetadata.js';
+import { allBlogs } from '../.contentlayer/generated/index.mjs';
 
 export async function getAllTags() {
-  const tagCount = {}
+  const tagCount = {};
   // Iterate through each post, putting all found tags into `tags`
   allBlogs.forEach((file) => {
     if (file.tags && file.draft !== true) {
       file.tags.forEach((tag) => {
-        const formattedTag = GithubSlugger.slug(tag)
+        const formattedTag = GithubSlugger.slug(tag);
         if (formattedTag in tagCount) {
-          tagCount[formattedTag] += 1
+          tagCount[formattedTag] += 1;
         } else {
-          tagCount[formattedTag] = 1
+          tagCount[formattedTag] = 1;
         }
-      })
+      });
     }
-  })
+  });
 
-  return tagCount
+  return tagCount;
 }
 
 const generateRssItem = (post) => `
@@ -34,7 +34,7 @@ const generateRssItem = (post) => `
     <author>${siteMetadata.email} (${siteMetadata.author})</author>
     ${post.tags && post.tags.map((t) => `<category>${t}</category>`).join('')}
   </item>
-`
+`;
 
 const generateRss = (posts, page = 'feed.xml') => `
   <rss version="2.0" xmlns:atom="http://www.w3.org/2005/Atom">
@@ -50,29 +50,29 @@ const generateRss = (posts, page = 'feed.xml') => `
       ${posts.map(generateRssItem).join('')}
     </channel>
   </rss>
-`
+`;
 
 async function generate() {
   // RSS for blog post
   if (allBlogs.length > 0) {
-    const rss = generateRss(allBlogs)
-    writeFileSync('./public/feed.xml', rss)
+    const rss = generateRss(allBlogs);
+    writeFileSync('./public/feed.xml', rss);
   }
 
   // RSS for tags
   // TODO: use AllTags from contentlayer when computed docs is ready
   if (allBlogs.length > 0) {
-    const tags = await getAllTags()
+    const tags = await getAllTags();
     for (const tag of Object.keys(tags)) {
       const filteredPosts = allBlogs.filter(
         (post) => post.draft !== true && post.tags.map((t) => GithubSlugger.slug(t)).includes(tag)
-      )
-      const rss = generateRss(filteredPosts, `tags/${tag}/feed.xml`)
-      const rssPath = path.join('public', 'tags', tag)
-      mkdirSync(rssPath, { recursive: true })
-      writeFileSync(path.join(rssPath, 'feed.xml'), rss)
+      );
+      const rss = generateRss(filteredPosts, `tags/${tag}/feed.xml`);
+      const rssPath = path.join('public', 'tags', tag);
+      mkdirSync(rssPath, { recursive: true });
+      writeFileSync(path.join(rssPath, 'feed.xml'), rss);
     }
   }
 }
 
-generate()
+generate();
