@@ -1,21 +1,21 @@
-import { writeFileSync } from 'fs'
-import globby from 'globby'
-import prettier from 'prettier'
-import siteMetadata from '../data/siteMetadata.js'
-import { allBlogs } from '../.contentlayer/generated/index.mjs'
+import { writeFileSync } from 'fs';
+import globby from 'globby';
+import prettier from 'prettier';
+import siteMetadata from '../data/siteMetadata.js';
+import { allBlogs } from '../.contentlayer/generated/index.mjs';
 
 async function generate() {
-  const prettierConfig = await prettier.resolveConfig('./.prettierrc.js')
+  const prettierConfig = await prettier.resolveConfig('./.prettierrc.js');
   const contentPages = allBlogs
     .map((x) => `/${x._raw.flattenedPath}`)
-    .filter((x) => !x.draft && !x.canonicalUrl)
+    .filter((x) => !x.draft && !x.canonicalUrl);
   const pages = await globby([
     'pages/*.{js|tsx}',
     'public/tags/**/*.xml',
     '!pages/_*.{js|tsx}',
     '!pages/api',
     '!pages/404.{js|tsx}',
-  ])
+  ]);
 
   const sitemap = `
         <?xml version="1.0" encoding="UTF-8"?>
@@ -29,24 +29,24 @@ async function generate() {
                   .replace('.js', '')
                   .replace('.mdx', '')
                   .replace('.md', '')
-                  .replace('/feed.xml', '')
-                const route = path === '/index' ? '' : path
+                  .replace('/feed.xml', '');
+                const route = path === '/index' ? '' : path;
                 return `
                         <url>
                             <loc>${siteMetadata.siteUrl}${route}</loc>
                         </url>
-                    `
+                    `;
               })
               .join('')}
         </urlset>
-    `
+    `;
 
   const formatted = prettier.format(sitemap, {
     ...prettierConfig,
     parser: 'html',
-  })
+  });
 
-  writeFileSync('public/sitemap.xml', formatted)
+  writeFileSync('public/sitemap.xml', formatted);
 }
 
-generate()
+generate();
