@@ -1,7 +1,7 @@
-const { replace } = ''
+const { replace } = '';
 
 // escape
-const ca = /[&<>'"]/g
+const ca = /[&<>'"]/g;
 
 const esca = {
   '&': '&amp;',
@@ -9,12 +9,12 @@ const esca = {
   '>': '&gt;',
   "'": '&#39;',
   '"': '&quot;',
-}
+};
 /**
  *
  * @param {keyof typeof esca} m
  */
-const pe = (m) => esca[m]
+const pe = (m) => esca[m];
 
 /**
  * Safely escape HTML entities such as `&`, `<`, `>`, `"`, and `'`.
@@ -23,4 +23,4 @@ const pe = (m) => esca[m]
  *  the input type is unexpected, except for boolean and numbers,
  *  converted as string.
  */
-export const escape = (es) => replace.call(es, ca, pe)
+export const escape = (es) => replace.call(es, ca, pe);
